Refactor interfaces to Application.Interfaces namespace
Move ICloudinaryService, IComicSearchService, and IOpenLibraryService (and related records) from Application.ComicInfoSearch to Application.Interfaces. Update all references and using statements accordingly. Add Regex.Match timeout in ComicSearchService for improved safety. No functional changes; this is an organizational refactor for better code structure.
