+++ Application/ComicInfoSearch/OpenLibraryService.cs
@@ -1,6 +1,7 @@
 using System.Net.Http.Json;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Application.Interfaces;
 
 namespace Application.ComicInfoSearch;
