+++ Persistence/Services/CloudinaryService.cs
@@ -1,4 +1,5 @@
 using Application.ComicInfoSearch;
+using Application.Interfaces;
 using CloudinaryDotNet;
 using CloudinaryDotNet.Actions;
 using Microsoft.Extensions.Options;
