+++ Web/Components/Pages/Books/AddBookForm.razor
@@ -2,6 +2,7 @@
 @rendermode InteractiveServer
 
 @using Application.ComicInfoSearch
+@using Application.Interfaces
 @using Microsoft.AspNetCore.Components
 @using Web.Services
 @using Web.Validators
