+++ tests/Application.UnitTests/ComicInfoSearch/ComicSearchServiceTests.cs
@@ -1,4 +1,5 @@
 using Application.ComicInfoSearch;
+using Application.Interfaces;
 using Microsoft.Extensions.Options;
 using NSubstitute;
 using NSubstitute.ExceptionExtensions;
